requirements.txt
python-telegram-bot==20.3
requests==2.31.0
urllib3==2.0.7

